requirements.txt
boto3==1.8.2
python-json-logger==v0.1.4
certifi==2020.4.5.1
python-decouple==3.3
Flask==1.0.2
json2html==1.3.0